requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
Flask==3.0.3
Flask-Cors==4.0.1
frozenlist==1.4.1
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
multidict==6.0.5
numpy==1.26.4
openai==0.28.0
packaging==24.1
pandas==2.2.1
pillow==10.3.0
pydantic==2.8.2
pydantic_core==2.20.1
pygame==2.5.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
scipy==1.12.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
tabulate==0.9.0
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.3
xlwt==1.3.0
yarl==1.9.4
python-dotenv